Add JIT highlighting
